# ci: use @swc/jest to speedup tests