--- a/jest.config.js
+++ b/jest.config.js
@@ -5,7 +5,9 @@ module.exports = {
     'lcov',
     'text'
   ],
-  preset: 'ts-jest',
+  transform: {
+    '^.+\\.(t|j)sx?$': '@swc/jest',
+  },
   // The root of your source code, typically /src
   // `<rootDir>` is a token Jest substitutes
   roots: ['<rootDir>'],