--- a/package.json
+++ b/package.json
@@ -37,6 +37,8 @@
     "@apidevtools/json-schema-ref-parser": "^9.0.9",
     "@apidevtools/swagger-parser": "^10.0.3",
     "@asyncapi/parser": "^2.0.0-next-major.5",
+    "@swc/core": "^1.3.5",
+    "@swc/jest": "^0.2.23",
     "alterschema": "^1.1.1",
     "change-case": "^4.1.2",
     "openapi-types": "9.3.0",