<!--   Thank you for your contribution. Before you submit the pull request:
1. Follow our contribution guidelines
2. Test your changes and attach their results to the pull request.
3. Update the relevant documentation.
-->

**Description**

Use `@swc/jest` to speedup tests.

`test:library`:
  - current `ts-jest` preset: 64.586s
  - `@swc/jest`: 36.319s
  
`test:examples`:
  - current `ts-jest` preset: 48.714s
  - `@swc/jest`: 27.995s